refactor: attachment, orguser, orgtoken, appsettings and projectuser
