+++ packages/nocodb/src/lib/controllers/orgUserController.ts
@@ -1,36 +1,16 @@
-import { Router } from 'express'
-import {
-  AuditOperationSubTypes,
-  AuditOperationTypes,
-  PluginCategory,
-} from 'nocodb-sdk'
-import { v4 as uuidv4 } from 'uuid'
-import validator from 'validator'
-import { OrgUserRoles } from 'nocodb-sdk'
-import { NC_APP_SETTINGS } from '../constants'
-import Audit from '../models/Audit'
-import ProjectUser from '../models/ProjectUser'
-import Store from '../models/Store'
-import SyncSource from '../models/SyncSource'
-import User from '../models/User'
-import Noco from '../Noco'
-import { MetaTable } from '../utils/globals'
-import { Tele } from 'nc-help'
-import { metaApiMetrics } from '../meta/helpers/apiMetrics'
-import { NcError } from '../meta/helpers/catchError'
-import { extractProps } from '../meta/helpers/extractProps'
-import ncMetaAclMw from '../meta/helpers/ncMetaAclMw'
-import { randomTokenString } from '../meta/helpers/stringHelpers'
-import { getAjvValidatorMw } from '../meta/api/helpers'
-import { sendInviteEmail } from '../meta/api/projectUserApis'
-import { orgUserService } from '../services'
+import { Router } from 'express';
+import { OrgUserRoles } from 'nocodb-sdk';
+import { metaApiMetrics } from '../meta/helpers/apiMetrics';
+import ncMetaAclMw from '../meta/helpers/ncMetaAclMw';
+import { getAjvValidatorMw } from '../meta/api/helpers';
+import { orgUserService } from '../services';
 
 async function userList(req, res) {
   res.json(
     await orgUserService.userList({
       query: req.query,
-    }),
-  )
+    })
+  );
 }
 
 async function userUpdate(req, res) {
@@ -38,72 +18,72 @@ async function userUpdate(req, res) {
     await orgUserService.userUpdate({
       user: req.body,
       userId: req.params.userId,
-    }),
-  )
+    })
+  );
 }
 
 async function userDelete(req, res) {
   await orgUserService.userDelete({
     userId: req.params.userId,
-  })
-  res.json({ msg: 'success' })
+  });
+  res.json({ msg: 'success' });
 }
 
-async function userAdd(req, res, next) {
+async function userAdd(req, res) {
   const result = await orgUserService.userAdd({
     user: req.body,
     req,
     projectId: req.params.projectId,
-  })
+  });
 
-  res.json(result)
+  res.json(result);
 }
 
 async function userSettings(_req, res): Promise<any> {
-  await orgUserService.userSettings({})
-  res.json({})
+  await orgUserService.userSettings({});
+  res.json({});
 }
 
 async function userInviteResend(req, res): Promise<any> {
   await orgUserService.userInviteResend({
     userId: req.params.userId,
     req,
-  })
+  });
 
-  res.json({ msg: 'success' })
+  res.json({ msg: 'success' });
 }
 
 async function generateResetUrl(req, res) {
   const result = await orgUserService.generateResetUrl({
     siteUrl: req.ncSiteUrl,
     userId: req.params.userId,
-  })
+  });
 
-  res.json(result)
+  res.json(result);
 }
 
 async function appSettingsGet(_req, res) {
-  const settings = await orgUserService.appSettingsGet()
-  res.json(settings)
+  const settings = await orgUserService.appSettingsGet();
+  res.json(settings);
 }
 
 async function appSettingsSet(req, res) {
   await orgUserService.appSettingsSet({
     settings: req.body,
-  })
+  });
 
-  res.json({ msg: 'Settings saved' })
+  res.json({ msg: 'Settings saved' });
 }
 
-const router = Router({ mergeParams: true })
+const router = Router({ mergeParams: true });
 router.get(
   '/api/v1/users',
   metaApiMetrics,
   ncMetaAclMw(userList, 'userList', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 router.patch(
   '/api/v1/users/:userId',
   metaApiMetrics,
@@ -111,16 +91,16 @@ router.patch(
   ncMetaAclMw(userUpdate, 'userUpdate', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 router.delete(
   '/api/v1/users/:userId',
   metaApiMetrics,
   ncMetaAclMw(userDelete, 'userDelete', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 router.post(
   '/api/v1/users',
   metaApiMetrics,
@@ -128,24 +108,24 @@ router.post(
   ncMetaAclMw(userAdd, 'userAdd', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 router.post(
   '/api/v1/users/settings',
   metaApiMetrics,
   ncMetaAclMw(userSettings, 'userSettings', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 router.post(
   '/api/v1/users/:userId/resend-invite',
   metaApiMetrics,
   ncMetaAclMw(userInviteResend, 'userInviteResend', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 
 router.post(
   '/api/v1/users/:userId/generate-reset-url',
@@ -153,8 +133,8 @@ router.post(
   ncMetaAclMw(generateResetUrl, 'generateResetUrl', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 
 router.get(
   '/api/v1/app-settings',
@@ -162,8 +142,8 @@ router.get(
   ncMetaAclMw(appSettingsGet, 'appSettingsGet', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 
 router.post(
   '/api/v1/app-settings',
@@ -171,7 +151,7 @@ router.post(
   ncMetaAclMw(appSettingsSet, 'appSettingsSet', {
     allowedRoles: [OrgUserRoles.SUPER_ADMIN],
     blockApiTokenAccess: true,
-  }),
-)
+  })
+);
 
-export default router
+export default router;
